Extract UseCase
